1. Move gridlayout.py from tests to phare/core    - Update usage of file in tests 2. Add Smallest patch size validation to simulation.py    - Using gridlayout.py to determine max ghosts 3  Move phare.init from src/phare to root/phare    - Move populate dict to phare/pharein
