+++ phare/init.py
@@ -1,9 +1,7 @@
-#! /usr/bin/env python3
-
 
 def get_user_inputs(jobname):
-    from phare import populateDict
     import importlib
+    from phare.pharein import populateDict
     print(jobname)
     jobmodule = importlib.import_module(jobname)
-    populateDict() # in __init__.py
+    populateDict()
