+++ src/phare/CMakeLists.txt
@@ -12,12 +12,3 @@ target_include_directories(${PROJECT_NAME} PUBLIC
 
 target_link_libraries(phare-exe PUBLIC phare_core phare_amr phare_initializer phare_solver phare_simulator)
 
-
-# this is the script that PHARE runs to prepare pythonpath in case pyphare is
-# ran in a virtualenv
-configure_file(${CMAKE_CURRENT_SOURCE_DIR}/init.py ${CMAKE_BINARY_DIR}/phare/init.py @ONLY)
-configure_file(${CMAKE_CURRENT_SOURCE_DIR}/__init__.py ${CMAKE_BINARY_DIR}/phare/__init__.py @ONLY)
-configure_file(${CMAKE_CURRENT_SOURCE_DIR}/phare_init.py ${CMAKE_CURRENT_BINARY_DIR}/phare_init.py @ONLY)
-
-
-
