+++ tests/core/data/gridlayout/deriv.py
@@ -7,7 +7,7 @@ import scipy.misc
 
 import sys
 
-import gridlayout
+from phare.core import gridlayout
 import os
 import gridparams
 import cellCenteredCoordinates
