+++ tests/core/data/gridlayout/gridIndexing.py
@@ -10,7 +10,7 @@ import numpy as np
 
 import sys
 
-import gridlayout
+from phare.core import gridlayout
 import os
 import utilities
 import gridparams
