+++ tests/core/data/gridlayout/test_laplacian.py
@@ -6,10 +6,7 @@
 import os
 import sys
 import numpy as np
-
-sys.path.append(os.path.join(os.path.dirname(__file__), '../../data/', 'gridlayout'))
-
-import gridlayout
+from phare.core import gridlayout
 
 
 class TestVariables (object) :
