+++ tests/core/numerics/ampere/test_ampere.py
@@ -7,11 +7,10 @@
 import sys
 import os
 
-sys.path.append(os.path.join(os.path.dirname(__file__), '../../data/', 'gridlayout'))
 
 
 import numpy as np
-import gridlayout
+from phare.core import gridlayout
 
 class TestVariables (object) :
