+++ tests/core/numerics/ohm/test_ohm.py
@@ -7,9 +7,8 @@
 import os
 import sys
 import numpy as np
-sys.path.append(os.path.join(os.path.dirname(__file__), '../../data/', 'gridlayout'))
 
-import gridlayout
+from phare.core import gridlayout
 
 class TestVariables (object) :
